app/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
Flask-Cors==4.0.0
